// src/types.ts
export interface NavItem {
  title: string;
  path: string;
}

export interface ServiceItem {
  title: string;
  description: string;
  icon: string;
}

export interface StepItem {
  number: string;
  title: string;
  description: string;
}

export interface FAQItem {
  question: string;
  answer: string;
}